+++ bin/miden-cli/src/lib.rs
@@ -10,10 +10,8 @@ use miden_client::{
         sqlite_store::SqliteStore, NoteFilter as ClientNoteFilter, OutputNoteRecord, Store,
         StoreAuthenticator,
     },
-    transactions::{LocalTransactionProver, TransactionProver},
     Client, ClientError, Felt, IdPrefixFetchError,
 };
-use miden_tx_prover::RemoteTransactionProver;
 use rand::Rng;
 mod commands;
 use commands::{
@@ -110,17 +108,11 @@ impl Cli {
         let rng = RpoRandomCoin::new(coin_seed.map(Felt::new));
         let authenticator = StoreAuthenticator::new_with_rng(store.clone() as Arc<dyn Store>, rng);
 
-        let tx_prover: Arc<dyn TransactionProver> = match &cli_config.remote_prover_endpoint {
-            Some(proving_url) => Arc::new(RemoteTransactionProver::new(&proving_url.to_string())),
-            None => Arc::new(LocalTransactionProver::new(Default::default())),
-        };
-
         let client = Client::new(
             Box::new(TonicRpcClient::new(&cli_config.rpc)),
             rng,
             store as Arc<dyn Store>,
             Arc::new(authenticator),
-            tx_prover as Arc<dyn TransactionProver>,
             in_debug_mode,
         );