+++ crates/rust-client/src/mock.rs
@@ -25,7 +25,7 @@ use miden_objects::{
     transaction::{InputNote, ProvenTransaction},
     BlockHeader, Digest, Felt, Word,
 };
-use miden_tx::{testing::mock_chain::MockChain, LocalTransactionProver};
+use miden_tx::testing::mock_chain::MockChain;
 use rand::Rng;
 use tonic::Response;
 use uuid::Uuid;
@@ -344,9 +344,7 @@ pub async fn create_test_client() -> (MockClient, MockRpcApi) {
     let rpc_api = MockRpcApi::new();
     let boxed_rpc_api = Box::new(rpc_api.clone());
 
-    let prover = Arc::new(LocalTransactionProver::default());
-
-    let client = MockClient::new(boxed_rpc_api, rng, store, Arc::new(authenticator), prover, true);
+    let client = MockClient::new(boxed_rpc_api, rng, store, Arc::new(authenticator), true);
     (client, rpc_api)
 }