+++ crates/web-client/src/lib.rs
@@ -5,7 +5,6 @@ use console_error_panic_hook::set_once;
 use miden_client::{
     rpc::WebTonicRpcClient,
     store::{web_store::WebStore, StoreAuthenticator},
-    transactions::{LocalTransactionProver, TransactionProver},
     Client,
 };
 use miden_objects::{crypto::rand::RpoRandomCoin, Felt};
@@ -27,6 +26,7 @@ pub mod transactions;
 #[wasm_bindgen]
 pub struct WebClient {
     store: Option<Arc<WebStore>>,
+    remote_prover: Option<Arc<RemoteTransactionProver>>,
     inner: Option<Client<RpoRandomCoin>>,
 }
 
@@ -41,7 +41,11 @@ impl WebClient {
     #[wasm_bindgen(constructor)]
     pub fn new() -> Self {
         set_once();
-        WebClient { inner: None, store: None }
+        WebClient {
+            inner: None,
+            remote_prover: None,
+            store: None,
+        }
     }
 
     pub(crate) fn get_mut_inner(&mut self) -> Option<&mut Client<RpoRandomCoin>> {
@@ -66,19 +70,10 @@ impl WebClient {
             &node_url.unwrap_or_else(|| "http://18.203.155.106:57291".to_string()),
         ));
 
-        let tx_prover: Arc<dyn TransactionProver> = match proving_url {
-            Some(proving_url) => Arc::new(RemoteTransactionProver::new(&proving_url.to_string())),
-            None => Arc::new(LocalTransactionProver::new(Default::default())),
-        };
-
-        self.inner = Some(Client::new(
-            web_rpc_client,
-            rng,
-            web_store.clone(),
-            authenticator,
-            tx_prover,
-            false,
-        ));
+        self.remote_prover = proving_url
+            .map(|proving_url| Arc::new(RemoteTransactionProver::new(&proving_url.to_string())));
+        self.inner =
+            Some(Client::new(web_rpc_client, rng, web_store.clone(), authenticator, false));
         self.store = Some(web_store);
 
         Ok(JsValue::from_str("Client created successfully"))